Improvements and fixes - Improve cli for Node.js management. - Implement basic cli for ADONE management.

+++ src/commands/adone/index.js
@@ -1,28 +1,125 @@
+import ADONEManager from "./manager";
+
 const {
     cli,
+    is,
     error,
     fs,
     app: {
         Subsystem,
         command
     },
-    path: aPath
+    path: aPath,
+    pretty
 } = adone;
 const { chalk, style, chalkify } = cli;
 
 const versionRegex = /^v\d+\.\d+\.\d+/;
 
+const activeStyle = chalkify("bold.underline", chalk);
+const cachedStyle = chalkify("#388E3C", chalk);
+const inactiveStyle = chalkify("white", chalk);
+const bullet = `${adone.text.unicode.symbol.bullet} `;
+const indent = " ".repeat(bullet.length);
+
+const getCurrentVersion = async () => {
+    try {
+        const exePath = await fs.which("adone");
+        return adone.process.execStdout(exePath, ["--version"]);
+    } catch (err) {
+        return "";
+    }
+};
+
 export default class ADONECommand extends Subsystem {
     onConfigure() {
+        this.adoneManager = new ADONEManager({
+            realm: kri.realm
+        });
+
         this.log = this.root.log;
     }
 
     @command({
         name: ["list", "ls"],
-        description: "Show ADONE releases"
+        description: "Show ADONE releases",
+        options: [
+            {
+                name: ["--date", "-D"],
+                description: "Show release date"
+            }
+        ]
     })
-    list() {
+    async list(args, opts) {
+        try {
+            const options = opts.getAll();
 
+            this.log({
+                message: "collecting release information"
+            });
+            const releases = await this.adoneManager.getVersions();
+            const currentVersion = await getCurrentVersion();
+            const downloadedVersions = await this.adoneManager.getDownloadedVersions();
+
+            const styledItem = (item) => {
+                let result = inactiveStyle(item.version);
+                const isCurrent = item.version === currentVersion;
+
+                if (isCurrent) {
+                    result = `${bullet}${`${activeStyle(item.version)}`}`;
+                } else {
+                    result = `${indent}${item.version}`;
+                }
+
+                if (downloadedVersions.includes(item.version)) {
+                    result = cachedStyle(result);
+                }
+                return result;
+            };
+
+            const model = [
+                {
+                    id: "version",
+                    handle: (item) => `${styledItem(item)}${item.lts ? chalk.grey(" (LTS)") : ""}`
+                }
+            ];
+
+            if (options.date) {
+                model.push({
+                    id: "date",
+                    width: 12,
+                    align: "right",
+                    handle: (item) => chalk.grey(item.date)
+                });
+            }
+
+            this.log({
+                message: "done",
+                clean: true,
+                status: true
+            });
+
+            console.log(pretty.table(releases, {
+                borderless: true,
+                noHeader: true,
+                style: {
+                    head: null,
+                    "padding-left": 1,
+                    compact: true
+                },
+                model
+            }));
+
+            return 0;
+        } catch (err) {
+            this.log({
+                message: err.message,
+                status: false,
+                // clean: true
+            });
+            // console.log(pretty.error(err));
+            return 1;
+        }
     }
 
     @command({
@@ -52,44 +149,69 @@ export default class ADONECommand extends Subsystem {
                 default: "adone",
                 description: "Global module name associated with spawned realm"
             },
+            // {
+            //     name: ["--bin-name", "-B"],
+            //     type: String,
+            //     default: "adone",
+            //     description: "Global executable name of ADONE cli"
+            // },
             {
-                name: ["--bin-name", "-B"],
+                name: "--node-version",
                 type: String,
-                default: "adone",
-                description: "Global executable name of ADONE cli"
+                default: process.version.split(".")[0].slice(1),
+                description: "Node.js version"
             }
         ]
     })
     async spawn(args, opts) {
-        const { source, path, dirName } = opts.getAll();
+        const { source, nodeVersion, path, dirName, moduleName } = opts.getAll();
 
         const destPath = aPath.resolve(path, dirName);
 
+        let canUndo = false;
         try {
             if (await fs.pathExists(destPath)) {
                 throw new error.ExistsException(`Path '${destPath}' already exists`);
             }
 
-            await fs.mkdirp(destPath);
+            canUndo = true;
 
-            if (source === "kri") {
+            let adoneSrcPath;
+            if (source === "latest" || versionRegex.test(source)) {
+                const relInfo = {
+                    version: source,
+                    nodeVersion
+                };
+                await this.adoneManager.download({
+                    ...relInfo,
+                    progressBar: true
+                });
+
+                this.log({
+                    message: "extracting files"
+                });
+                adoneSrcPath = await this.adoneManager.extract(relInfo);
+            } else if (source === "kri") {
                 // extract bundled ADONE realm
 
                 this.log({
                     message: "extracting files from KRI"
                 });
 
-                await fs.copyEx("/adone", destPath, {
-                    results: false,
-                    dot: true,
-                    junk: true
-                });
-            } else if (versionRegex.test(source)) {
-                //
+                adoneSrcPath = "/adone";
             }
 
+            await fs.copyEx(adoneSrcPath, destPath, {
+                results: false,
+                dot: true,
+                junk: true
+            });
+
+            const modulePath = aPath.join(adone.system.env.home(), ".node_modules", moduleName);
+            await this._createSymlink(destPath, modulePath);
+
             this.log({
-                message: "done",
+                message: `note: to run ADONE/cli from anywhere add ${style.primary(aPath.join(destPath, "bin"))} to PATH variable\ndone`,
                 status: true
             });
 
@@ -102,8 +224,19 @@ export default class ADONECommand extends Subsystem {
             });
             console.error(adone.pretty.error(err));
 
-            await fs.remove(destPath);
+            if (canUndo) {
+                await fs.remove(destPath);
+            }
             return 1;
+        }
+    }
+
+    async _createSymlink(targetPath, path) {
+        await fs.mkdirp(aPath.dirname(path));
+        if (is.windows) {
+            await fs.symlink(targetPath, path, "junction");
+        } else {
+            await fs.symlink(targetPath, path);
         }
     }
 }
